Editor: Creating a BlockDropZone component to allow dropping files at a given position
