+++ editor/modes/visual-editor/block.js
@@ -4,16 +4,16 @@
 import { connect } from 'react-redux';
 import classnames from 'classnames';
 import { Slot } from 'react-slot-fill';
-import { partial, reduce, get, find } from 'lodash';
+import { partial } from 'lodash';
 import CSSTransitionGroup from 'react-transition-group/CSSTransitionGroup';
 
 /**
  * WordPress dependencies
  */
 import { Children, Component } from '@wordpress/element';
-import { IconButton, Toolbar, DropZone } from '@wordpress/components';
+import { IconButton, Toolbar } from '@wordpress/components';
 import { keycodes } from '@wordpress/utils';
-import { getBlockType, getBlockDefaultClassname, createBlock, getBlockTypes } from '@wordpress/blocks';
+import { getBlockType, getBlockDefaultClassname, createBlock } from '@wordpress/blocks';
 import { __, sprintf } from '@wordpress/i18n';
 
 /**
@@ -22,6 +22,7 @@ import { __, sprintf } from '@wordpress/i18n';
 import InvalidBlockWarning from './invalid-block-warning';
 import BlockCrashWarning from './block-crash-warning';
 import BlockCrashBoundary from './block-crash-boundary';
+import BlockDropZone from './block-drop-zone';
 import BlockMover from '../../block-mover';
 import BlockRightMenu from '../../block-settings-menu';
 import BlockSwitcher from '../../block-switcher';
@@ -75,7 +76,7 @@ class VisualEditorBlock extends Component {
 		this.onKeyUp = this.onKeyUp.bind( this );
 		this.toggleMobileControls = this.toggleMobileControls.bind( this );
 		this.onBlockError = this.onBlockError.bind( this );
-		this.dropFiles = this.dropFiles.bind( this );
+		this.insertBlocksAfter = this.insertBlocksAfter.bind( this );
 
 		this.previousOffset = null;
 
@@ -233,6 +234,10 @@ class VisualEditorBlock extends Component {
 		}
 	}
 
+	insertBlocksAfter( blocks ) {
+		this.props.onInsertBlocks( blocks, this.props.order + 1 );
+	}
+
 	onFocus( event ) {
 		if ( event.target === this.node ) {
 			this.props.onSelect();
@@ -275,25 +280,6 @@ class VisualEditorBlock extends Component {
 		this.setState( { error } );
 	}
 
-	dropFiles( files, position ) {
-		const { order } = this.props;
-		const transformation = reduce( getBlockTypes(), ( ret, blockType ) => {
-			if ( ret ) {
-				return ret;
-			}
-
-			return find( get( blockType, 'transforms.from', [] ), ( transform ) => (
-				transform.type === 'files' && transform.isMatch( files )
-			) );
-		}, false );
-
-		if ( transformation ) {
-			transformation.transform( files ).then( ( blocks ) => {
-				this.props.onInsertBlocks( blocks, position.y === 'top' ? order : order + 1 );
-			} );
-		}
-	}
-
 	render() {
 		const { block, multiSelectedBlockUids, order } = this.props;
 		const { name: blockName, isValid } = block;
@@ -328,7 +314,7 @@ class VisualEditorBlock extends Component {
 			'is-showing-mobile-controls': showMobileControls,
 		} );
 
-		const { onMouseLeave, onFocus, onInsertBlocks, onReplace } = this.props;
+		const { onMouseLeave, onFocus, onReplace } = this.props;
 
 		// Determine whether the block has props to apply to the wrapper.
 		let wrapperProps;
@@ -357,9 +343,7 @@ class VisualEditorBlock extends Component {
 				aria-label={ blockLabel }
 				{ ...wrapperProps }
 			>
-				<DropZone
-					onFilesDrop={ this.dropFiles }
-				/>
+				<BlockDropZone index={ order } />
 				{ ( showUI || isHovered ) && <BlockMover uids={ [ block.uid ] } /> }
 				{ ( showUI || isHovered ) && <BlockRightMenu uid={ block.uid } /> }
 				{ showUI && isValid &&
@@ -406,7 +390,7 @@ class VisualEditorBlock extends Component {
 								focus={ focus }
 								attributes={ block.attributes }
 								setAttributes={ this.setAttributes }
-								insertBlocksAfter={ ( blocks ) => onInsertBlocks( blocks, order + 1 ) }
+								insertBlocksAfter={ this.insertBlocksAfter }
 								onReplace={ onReplace }
 								setFocus={ partial( onFocus, block.uid ) }
 								mergeBlocks={ this.mergeBlocks }
